update starter card carousel
