+++ packages/scoutgame-ui/src/components/common/Card/BuilderCard/BuilderCard.tsx
@@ -3,6 +3,7 @@
 import { Card, Stack, Typography } from '@mui/material';
 import type { BuilderInfo } from '@packages/scoutgame/builders/interfaces';
 
+import { useLgScreen, useMdScreen } from '../../../../hooks/useMediaScreens';
 import { ScoutButton } from '../../ScoutButton/ScoutButton';
 
 import { BuilderCardNftDisplay } from './BuilderCardNftDisplay';
@@ -14,7 +15,7 @@ export function BuilderCard({
   builder,
   showPurchaseButton = false,
   hideDetails = false,
-  size = 'medium',
+  size: sizeOverride,
   disableProfileUrl = false,
   markStarterCardPurchased = false,
   type
@@ -27,6 +28,9 @@ export function BuilderCard({
   markStarterCardPurchased?: boolean;
   type: 'default' | 'starter_pack';
 }) {
+  const isDesktop = useMdScreen();
+  const isLgScreen = useLgScreen();
+  const size = sizeOverride || (isLgScreen ? 'large' : isDesktop ? 'small' : 'x-small');
   return (
     <Card
       sx={{
